Use static beans.xml files
