+++ dev/fattest.simplicity/bnd.bnd
@@ -18,6 +18,7 @@ Bundle-Description: FAT infrastructure; version=${bVersion}
 Export-Package: \
 	com.ibm.websphere.simplicity;version=1.0.16,\
 	com.ibm.websphere.simplicity.application;version=1.0.16,\
+	com.ibm.websphere.simplicity.beansxml;version=1.0.16,\
 	com.ibm.websphere.simplicity.config;version=1.0.16,\
 	com.ibm.websphere.simplicity.config.context;version=1.0.16,\
 	com.ibm.websphere.simplicity.config.dsprops;version=1.0.16,\
